+++ test/Telegram.Bot.Tests.Integ/Admin Bot/ChannelAdminBotTests.cs
@@ -177,11 +177,11 @@ namespace Telegram.Bot.Tests.Integ.Admin_Bot
         {
             await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldThrowOnDeletingChatDeletedPhoto);
 
-            Exception e = await Assert.ThrowsAnyAsync<Exception>(() =>
+            ApiRequestException exception = await Assert.ThrowsAnyAsync<ApiRequestException>(() =>
                 BotClient.DeleteChatPhotoAsync(_classFixture.ChatId));
 
-            Assert.IsType<ApiRequestException>(e);
-            Assert.Equal("Bad Request: CHAT_NOT_MODIFIED", e.Message);
+            Assert.IsType<BadRequestException>(exception);
+            Assert.Equal("CHAT_NOT_MODIFIED", exception.Message);
         }
 
         #endregion
@@ -202,8 +202,8 @@ namespace Telegram.Bot.Tests.Integ.Admin_Bot
             );
 
             // ToDo: Create exception type
-            Assert.Equal(400, exception.ErrorCode);
-            Assert.Equal("Bad Request: method is available only for supergroups", exception.Message);
+            Assert.IsType<BadRequestException>(exception);
+            Assert.Equal("method is available only for supergroups", exception.Message);
         }
 
         #endregion